fix type

+++ src/xtd.core/include/xtd/ticks.h
@@ -5,6 +5,7 @@
 
 #include <chrono>
 #include "tick.h"
+#include "types.h"
 
 /// @brief The xtd namespace contains all fundamental classes to access Hardware, Os, System, and more.
 namespace xtd {
@@ -14,5 +15,5 @@ namespace xtd {
   /// @par Library
   /// xtd.core
   /// @ingroup xtd_core types
-  using ticks = std::chrono::duration<long long, tick>;
+  using ticks = std::chrono::duration<int64, tick>;
 }
